Seperate main view model to individual view models
