+++ src/MainWindowViewModel.cs
@@ -1,36 +1,14 @@
 ﻿using System.ComponentModel;
-using System.IO;
 using System.Runtime.CompilerServices;
-using System.Text;
-using Microsoft.Win32;
-using Newtonsoft.Json;
 using RaceResultConverter.Zon;
 
 namespace RaceResultConverter;
 
 public class MainWindowViewModel : INotifyPropertyChanged
 {
-    private bool _convertResult;
-    private string _convertMessage = string.Empty;
-    private string _selectedFilePath = string.Empty;
+
     private ConvertType _convertType;
-    private readonly OpenFileDialog _openFileDialog;
     private ConvertSourceType _convertSourceType;
-
-    public RelayCommand ConvertCommand { get; set; }
-
-    public RelayCommand SelectFileCommand { get; set; }
-
-    public string SelectedFilePath
-    {
-        get => _selectedFilePath;
-        set
-        {
-            _selectedFilePath = value;
-            OnPropertyChanged(nameof(SelectedFilePath));
-            ConvertCommand.RaiseCanExecuteChanged();
-        }
-    }
 
     public ConvertType ConvertType
     {
@@ -38,35 +16,10 @@ public class MainWindowViewModel : INotifyPropertyChanged
         set
         {
             _convertType = value;
-            _openFileDialog.Filter = GetFileFilterString();
-            SelectedFilePath = string.Empty;
-
             OnPropertyChanged(nameof(ConvertType));
         }
     }
-
-    public string ConvertMessage
-    {
-        get => _convertMessage;
-        set
-        {
-            _convertMessage = value;
-            OnPropertyChanged(nameof(ConvertMessage));
-        }
-    }
-
-    public bool ConvertResult
-    {
-        get => _convertResult;
-        set
-        {
-            _convertResult = value;
-            OnPropertyChanged(nameof(ConvertResult));
-        }
-    }
-
-    private bool SourceFileSelected => !string.IsNullOrEmpty(SelectedFilePath) && File.Exists(SelectedFilePath);
-
+  
     public ConvertSourceType ConvertSourceType
     {
         get => _convertSourceType;
@@ -82,101 +35,13 @@ public class MainWindowViewModel : INotifyPropertyChanged
     public bool IsSingleConvert => ConvertSourceType == ConvertSourceType.Single;
     public bool IsChampionshipConvert => ConvertSourceType == ConvertSourceType.Championship;
 
+    public SingleRaceConvertViewModel SingleRaceConvertViewModel { get; }
+    public ChampionshipResultConvertViewModel ChampionshipResultConvertViewModel{ get; }
+
     public MainWindowViewModel()
     {
-        _openFileDialog = new OpenFileDialog
-        {
-            Filter = GetFileFilterString()
-        };
-
-        ConvertCommand = new RelayCommand(Convert, CanConvert);
-        SelectFileCommand = new RelayCommand(SelectFile);
-    }
-
-    private string GetFileFilterString()
-    {
-        return ConvertType == ConvertType.ZRoundToZon ? "ZRound Race File|*.rcf" : "JSON File|*.json";
-    }
-
-    private bool CanConvert() => SourceFileSelected && ConvertType == ConvertType.ZRoundToZon;    //Currently, only support convert ZRound result to Zon's.
-
-    private void Convert()
-    {
-        var result = InnerConvert(ConvertType, _openFileDialog.FileName);
-
-        var saveFileDialog = new SaveFileDialog
-        {
-            AddExtension = true,
-            CreatePrompt = true,
-            FileName = "Result",
-            DefaultExt = ".json",
-            Filter = "JSON File|*.json",
-            ValidateNames = true
-        };
-
-        if (saveFileDialog.ShowDialog() ?? false)
-        {
-            ConvertResult = SaveJson(result, saveFileDialog.FileName);
-            ConvertMessage = ConvertResult ? Resource.ConvertResult_Succeed : Resource.ConvertResult_Failed;
-
-            return;
-        }
-
-        ConvertMessage = Resource.ConvertResult_Cancelled;
-        ConvertResult = false;
-    }
-
-    private static IRaceResult InnerConvert(ConvertType convertType, string jsonFile)
-    {
-        return convertType switch
-        {
-            ConvertType.ZRoundToZon => new ZRoundResultConverter(ConvertZRoundToZon).ConvertToTarget(jsonFile),
-            ConvertType.ZonToZRound => new ResultConverter<ZonResult, ZRoundResult>(ConvertZonToZRound).ConvertToTarget(jsonFile),
-            _ => throw new ArgumentOutOfRangeException(nameof(convertType), convertType, null)
-        };
-    }
-
-    private void SelectFile()
-    {
-        if (_openFileDialog.ShowDialog() ?? false)
-            SelectedFilePath = _openFileDialog.FileName;
-    }
-
-    private static bool SaveJson(IRaceResult result, string outputPath)
-    {
-        try
-        {
-            if(result == null)
-                return false;
-
-            var s = JsonConvert.SerializeObject(result);
-            using var file = new FileStream(outputPath, FileMode.OpenOrCreate);
-            file.SetLength(0);
-            var bytes = Encoding.UTF8.GetBytes(s);
-            file.Write(bytes, 0, bytes.Length);
-
-            return true;
-        }
-        catch (Exception e)
-        {
-            Console.WriteLine(e);
-            return false;
-        }
-    }
-
-    private static ZonResult ConvertZRoundToZon(ZRoundResult arg)
-    {
-        return new ZonResult
-        {
-            RaceDataProp = new RaceDataProp
-            {
-                RaceTitle = arg.Name,
-                PrintTitle = arg.Name,
-                Time = arg.Start,
-                RaceRoundid = arg.Start,
-                RaceMode = 1
-            }
-        };
+        SingleRaceConvertViewModel = new SingleRaceConvertViewModel(() => ConvertType);
+        ChampionshipResultConvertViewModel = new ChampionshipResultConvertViewModel();
     }
 
     private static ZRoundResult ConvertZonToZRound(ZonResult arg)
