+++ src/ZRoundRaceResultItem.cs
@@ -1,0 +1,10 @@
+﻿namespace RaceResultConverter;
+
+public class ZRoundRaceResultItem
+{
+    public bool IsCheck { get; set; }
+
+    public string RcfFile { get; set; } = string.Empty;
+
+    public string JsonFile { get; set; } = string.Empty;
+}
